Always create drone file from scratch
The ensure_config function checks if an existing drone configuration
exists, reads it in, checks if the existing lines match the template,
and updates it, fixing mismatched template sections.  This is all
unnecessary since overwriting any existing file will have the same
effect and is much simpler.

I'm keeping the function around to use it for reading the existing
config, validating it, and loading the values into the form for
modification (future feature).